+++ python/sopActor/masterThread.py
@@ -5,7 +5,7 @@
 # @License: BSD 3-clause (http://www.opensource.org/licenses/BSD-3-Clause)
 #
 # @Last modified by: José Sánchez-Gallego (gallegoj@uw.edu)
-# @Last modified time: 2019-01-16 11:03:15
+# @Last modified time: 2019-01-16 14:51:50
 
 import Queue
 import threading
@@ -348,6 +348,10 @@ def prep_guider_decenter_off(multiCmd, precondition=True):
     Command: guider decenter off
     """
 
+    if myGlobals.bypass.get('guider_decenter'):
+        multiCmd.cmd.warn('text="skipping prep_guider_decenter_off '
+                          'because guider_decenter is bypassed."')
+
     if precondition:
         multiCmd.append(SopPrecondition(sopActor.GUIDER, Msg.DECENTER, on=False))
     else:
@@ -361,6 +365,11 @@ def prep_manga_dither(multiCmd, dither='C', precondition=False):
 
     Command: guider mangaDither ditherPos=N
     """
+
+    if myGlobals.bypass.get('guider_decenter'):
+        multiCmd.cmd.warn('text="skipping prep_manga_dither because '
+                          'guider_decenter is bypassed."')
+
     # append guider decenter on
     prep_guider_decenter_on(multiCmd)
     # append manga guider dither command